un-vibe coded tests

+++ src/zod/auth.ts
@@ -3,5 +3,5 @@ import z from "zod/v3";
 export const authSchema = z.object({
   id: z.string().uuid(),
   roles: z.array(z.string()),
-  iat: z.number(),
+  iat: z.number().int(),
 });
